fixes everywhere

+++ mahonyFilterMain.cpp
@@ -12,7 +12,7 @@ int main() {
     bool calculateBestK = false;
 
     const double dt = 0.02;
-    const double kp = 9; // Was calcualted via searching loop using RMSE for both roll and pitch
+    const double kp = 9; // Was calcualted via searching loop using RMSE for both roll and pitch best value is 11
     const int DISPLAY_SAMPLES = 10;
 
     // Load data
@@ -104,8 +104,8 @@ int main() {
     std::cout << "Pitch RMSE: " << Utils::rmse(pitchGroundTruth, mahony.getPitchEstimation()) << " degrees\n";
     std::cout << "Pitch MEA:  " << Utils::mea(pitchGroundTruth, mahony.getPitchEstimation()) << " degrees\n";
 
-    Utils::printVecToFile(mahony.getRollEstimation(), "Results/Results/MahonyRoll_kp_" + std::to_string(kp) + ".txt");
-    Utils::printVecToFile(mahony.getPitchEstimation(), "Results/Results/MahonyPitch_kp_" + std::to_string(kp) + ".txt");
+    Utils::printVecToFile(mahony.getRollEstimation(), "Results/Results/MahonyRoll_kp_" + std::to_string(static_cast<int>(kp)) + ".txt");
+    Utils::printVecToFile(mahony.getPitchEstimation(), "Results/Results/MahonyPitch_kp_" + std::to_string(static_cast<int>(kp)) + ".txt");
 
     return 0;
 }
